some cleaning up for #80
move get_torsion_info_angles to utilities
flesh out docstrings
linter cleanup
pytest cleanup
