+++ examples/torsion_example.py
@@ -2,59 +2,6 @@ import stk
 import stko
 
 
-def get_torsion_info_angles(mol, torsion_info):
-    torsion = torsion_info.get_torsion()
-    angle = stko.calculate_dihedral(
-        pt1=tuple(
-            mol.get_atomic_positions(
-                torsion.get_atom_ids()[0]
-            )
-        )[0],
-        pt2=tuple(
-            mol.get_atomic_positions(
-                torsion.get_atom_ids()[1]
-            )
-        )[0],
-        pt3=tuple(
-            mol.get_atomic_positions(
-                torsion.get_atom_ids()[2]
-            )
-        )[0],
-        pt4=tuple(
-            mol.get_atomic_positions(
-                torsion.get_atom_ids()[3]
-            )
-        )[0],
-    )
-    bb_torsion = torsion_info.get_building_block_torsion()
-    if bb_torsion is None:
-        bb_angle = None
-    else:
-        bb_angle = stko.calculate_dihedral(
-            pt1=tuple(
-                torsion_info.get_building_block().get_atomic_positions(
-                    bb_torsion.get_atom_ids()[0]
-                )
-            )[0],
-            pt2=tuple(
-                torsion_info.get_building_block().get_atomic_positions(
-                    bb_torsion.get_atom_ids()[1]
-                )
-            )[0],
-            pt3=tuple(
-                torsion_info.get_building_block().get_atomic_positions(
-                    bb_torsion.get_atom_ids()[2]
-                )
-            )[0],
-            pt4=tuple(
-                torsion_info.get_building_block().get_atomic_positions(
-                    bb_torsion.get_atom_ids()[3]
-                )
-            )[0],
-        )
-    return angle, bb_angle
-
-
 def main():
     bb1 = stk.BuildingBlock('NCCNCCN', [stk.PrimaryAminoFactory()])
     bb2 = stk.BuildingBlock('O=CCCC=O', [stk.AldehydeFactory()])
@@ -91,7 +38,7 @@ def main():
         )
     print(tors_results.get_torsion_infos_by_building_block())
     for t in tors_results.get_torsion_infos():
-        print(get_torsion_info_angles(polymer, t))
+        print(stko.get_torsion_info_angles(polymer, t))
 
 
 if __name__ == "__main__":
